feat: support webhook validate plugin (#1355)

+++ test/e2e/scaffold/scaffold.go
@@ -24,7 +24,6 @@ import (
 	"net/http"
 	"net/url"
 	"os"
-	"os/exec"
 	"os/user"
 	"path/filepath"
 	"regexp"
@@ -492,11 +491,6 @@ func (s *Scaffold) beforeEach() {
 
 	k8s.WaitUntilServiceAvailable(s.t, s.kubectlOptions, s.testBackendService.Name, 3, 2*time.Second)
 
-	if s.opts.EnableWebhooks {
-		err := generateWebhookCert(s.namespace)
-		assert.Nil(s.t, err, "generate certs and create webhook secret")
-	}
-
 	err = s.newIngressAPISIXController()
 	assert.Nil(s.t, err, "initializing ingress apisix controller")
 
@@ -510,8 +504,9 @@ func (s *Scaffold) afterEach() {
 	defer ginkgo.GinkgoRecover()
 
 	if ginkgo.CurrentSpecReport().Failed() {
-		if os.Getenv("E2E_ENV") == "ci" {
-			// dump and delete related resource
+		// dump and delete related resource
+		env := os.Getenv("E2E_ENV")
+		if env == "ci" || env == "debug" {
 			_, _ = fmt.Fprintln(ginkgo.GinkgoWriter, "Dumping namespace contents")
 			output, _ := k8s.RunKubectlAndGetOutputE(ginkgo.GinkgoT(), s.kubectlOptions, "get", "deploy,sts,svc,pods")
 			if output != "" {
@@ -531,6 +526,14 @@ func (s *Scaffold) afterEach() {
 			if output != "" {
 				_, _ = fmt.Fprintln(ginkgo.GinkgoWriter, output)
 			}
+			if s.opts.EnableWebhooks {
+				output, _ = k8s.RunKubectlAndGetOutputE(ginkgo.GinkgoT(), s.kubectlOptions, "get", "validatingwebhookconfigurations", "-o", "yaml")
+				if output != "" {
+					_, _ = fmt.Fprintln(ginkgo.GinkgoWriter, output)
+				}
+			}
+		}
+		if env != "debug" {
 			err := k8s.DeleteNamespaceE(s.t, s.kubectlOptions, s.namespace)
 			assert.Nilf(ginkgo.GinkgoT(), err, "deleting namespace %s", s.namespace)
 		}
@@ -648,19 +651,6 @@ func waitExponentialBackoff(condFunc func() (bool, error)) error {
 		Steps:    8,
 	}
 	return wait.ExponentialBackoff(backoff, condFunc)
-}
-
-// generateWebhookCert generates signed certs of webhook and create the corresponding secret by running a script.
-func generateWebhookCert(ns string) error {
-	commandTemplate := `testdata/webhook-create-signed-cert.sh`
-	cmd := exec.Command("/bin/sh", commandTemplate, "--namespace", ns)
-
-	output, err := cmd.Output()
-	if err != nil {
-		ginkgo.GinkgoT().Errorf("%s", output)
-		return fmt.Errorf("failed to execute the script: %v", err)
-	}
-	return nil
 }
 
 func (s *Scaffold) CreateVersionedApisixResource(yml string) error {
